Remove redundant cast

+++ src/listing.c
@@ -206,7 +206,7 @@ static void
 set_icon_names_hashes(void)
 {
 	size_t i = sizeof(icon_filenames) / sizeof(icon_filenames[0]);
-	name_icons_hashes = xnmalloc((size_t)i + 1, sizeof(size_t));
+	name_icons_hashes = xnmalloc(i + 1, sizeof(size_t));
 
 	for (; i-- > 0;)
 		name_icons_hashes[i] = hashme(icon_filenames[i].name, 0);
@@ -216,7 +216,7 @@ static void
 set_dir_names_hashes(void)
 {
 	size_t i = sizeof(icon_dirnames) / sizeof(icon_dirnames[0]);
-	dir_icons_hashes = xnmalloc((size_t)i + 1, sizeof(size_t));
+	dir_icons_hashes = xnmalloc(i + 1, sizeof(size_t));
 
 	for (; i-- > 0;)
 		dir_icons_hashes[i] = hashme(icon_dirnames[i].name, 0);
@@ -226,7 +226,7 @@ static void
 set_ext_names_hashes(void)
 {
 	size_t i = sizeof(icon_ext) / sizeof(icon_ext[0]);
-	ext_icons_hashes = xnmalloc((size_t)i + 1,  sizeof(size_t));
+	ext_icons_hashes = xnmalloc(i + 1,  sizeof(size_t));
 
 	for (; i-- > 0;)
 		ext_icons_hashes[i] = hashme(icon_ext[i].name, 0);
